// types/job-types.ts
export interface JobHighlights {
    Qualifications?: string[];
    Responsibilities?: string[];
  }
  
  export interface ApplyOption {
    publisher: string;
    apply_link: string;
    is_direct: boolean;
  }
  
  export interface Job {
    job_id: string;
    job_title: string;
    employer_name: string;
    employer_logo: string | null;
    employer_website: string | null;
    job_publisher: string;
    job_employment_type: string;
    job_employment_types: string[];
    job_apply_link: string;
    job_apply_is_direct: boolean;
    apply_options: ApplyOption[];
    job_description: string;
    job_is_remote: boolean;
    job_posted_at: string;
    job_posted_at_timestamp: number;
    job_posted_at_datetime_utc: string;
    job_location: string;
    job_city: string;
    job_state: string;
    job_country: string;
    job_latitude: number;
    job_longitude: number;
    job_benefits: string | null;
    job_google_link: string;
    job_salary: string | null;
    job_min_salary: number | null;
    job_max_salary: number | null;
    job_salary_period: string | null;
    job_highlights: JobHighlights;
    job_onet_soc: string;
    job_onet_job_zone: string;
  }
  
  export interface JobsResponse {
    status: string;
    request_id: string;
    parameters: {
      query?: string;
      page: number;
      num_pages: number;
      date_posted: string;
      country: string;
      language: string;
    };
    data: Job[];
  }
  
  export interface JobSearchFilters {
    query: string;
    country?: string;
    employmentType?: string;
    isRemote?: boolean;
    datePosted?: string;
  }

 export interface JobDetailsResponse extends Omit<JobsResponse, 'parameters'> {
    parameters: {
      job_id: string;
      country: string;
      language: string;
      page: number;
      num_pages: number;
      date_posted: string;
    };
  }